# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TravelAgent local API
  version: 0.1.1-design
  description: Design contract, not an implemented server. Loopback-only authenticated
    local application. Domain contracts now use NEED_LOGIN and VERIFICATION_REQUIRED;
    evidence includes source metadata, applicability and nullable confidence. PRIVATE_LOCAL_RESEARCH
    SourceContent and SourceBodyBlock are internal SQLite v5 contracts; v6 adds extraction
    attempts and safe diagnostics; v7 adds private per-candidate grounding/context
    review and partial-success states. v8 adds a single supplemental extraction authorization
    and transport-phase diagnostics. v10 extends the existing extraction authorizations
    and adds optional snapshot-bound span lineage and reviewed source roles. v9 adds
    a fixed research continuation ledger; optional ClaimAssessment route associations
    and historical author-trip scope remain private, source-anchored review metadata.
    These are not public HTTP response fields; no raw-content endpoint is added. P01
    /api/v1/preview routes are implemented; other business routes remain design contracts.
    SQLite v11 adds independent preview choices/receipts.
servers:
- url: http://127.0.0.1:8765
security:
- LocalSession: []
paths:
...
  /api/v1/preview/planning:
    get:
      operationId: readPlanningIndex
      summary: 日常私人工作台：新旅行无隐含授权；页面有限许可、追加/撤销复用耐久账本；本地复用/修改/恢复零外部调用；历史规则保留
      parameters: []
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/PlanIndex
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
    post:
      operationId: createIndependentTrip
      summary: 日常私人工作台：新旅行无隐含授权；页面有限许可、追加/撤销复用耐久账本；本地复用/修改/恢复零外部调用；历史规则保留
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/PlanView
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./domain.schema.json#/$defs/PlanCreate
  /api/v1/preview/planning/{session_id}:
    get:
      operationId: readPlanningDraft
      summary: 日常私人工作台：新旅行无隐含授权；页面有限许可、追加/撤销复用耐久账本；本地复用/修改/恢复零外部调用；历史规则保留
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/PlanView
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
    post:
      operationId: changePlanningDraft
      summary: 日常私人工作台：新旅行无隐含授权；页面有限许可、追加/撤销复用耐久账本；本地复用/修改/恢复零外部调用；历史规则保留
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/PlanView
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./domain.schema.json#/$defs/PlanAction
  /api/v1/preview/planning-maps/{session_id}:
    get:
      operationId: readPlanningPlaces
      summary: 日常私人工作台：新旅行无隐含授权；页面有限许可、追加/撤销复用耐久账本；本地复用/修改/恢复零外部调用；历史规则保留
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/MapView
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
  /api/v1/preview/planning-maps:
    post:
      operationId: changePlanningPlaces
      summary: 日常私人工作台：新旅行无隐含授权；页面有限许可、追加/撤销复用耐久账本；本地复用/修改/恢复零外部调用；历史规则保留
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/MapView
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./domain.schema.json#/$defs/MapAction
...
  /api/v1/preview/planning/{session_id}/guide-export:
    get:
      operationId: exportPlanningGuide
      summary: 日常私人工作台：新旅行无隐含授权；页面有限许可、追加/撤销复用耐久账本；本地复用/修改/恢复零外部调用；历史规则保留
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/GuideExport
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
  /api/v1/preview/automatic-planning:
    post:
      operationId: startAutomaticPlanning
      summary: 一次明确点击记录版本化私人用途与服务器有界上限，耐久编排研究/审核/建议；GET零派发，改条件隔离旧结果，最终采用仍显式确认
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/PlanView
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./domain.schema.json#/$defs/AutomaticStart
  /api/v1/preview/automatic-planning/{session_id}:
    post:
      operationId: changeAutomaticPlanning
      summary: 一次明确点击记录版本化私人用途与服务器有界上限，耐久编排研究/审核/建议；GET零派发，改条件隔离旧结果，最终采用仍显式确认
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/PlanView
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./domain.schema.json#/$defs/AutomaticAction
  /api/v1/preview/conversation/{session_id}:
    post:
      operationId: changePlanningConversation
      summary: 日常私人工作台：新旅行无隐含授权；页面有限许可、追加/撤销复用耐久账本；本地复用/修改/恢复零外部调用；历史规则保留
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/PlanView
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./domain.schema.json#/$defs/ConversationAction
  /api/v1/preview/key-leg/{session_id}:
    post:
      operationId: checkExplicitKeyLeg
      summary: 显式单段许可：公共地点最多2次、路径最多1次；实际POI需确认，无自动重试，刷新重启零派发
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/PlanView
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./domain.schema.json#/$defs/CriticalMapAction
  /api/v1/preview/planning/{session_id}/reference-overview-export:
    get:
      operationId: exportLocalRouteReference
      summary: 当前有效引用的版本化本地路线参考导出；不是已采用攻略或新模型生成，零外部调用
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      security:
      - PreviewMapSession: []
      responses:
        '200':
          description: 本机已审核缓存和用户选择
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ReferenceOverviewExport
        default:
          description: 安全错误；不返回原始异常或正文
          content:
            application/json:
              schema:
                $ref: ./domain.schema.json#/$defs/ErrorResponse
components:
  securitySchemes:
    LocalSession:
      type: apiKey
      in: cookie
      name: ta_session
    PreviewSession:
      type: apiKey
      in: cookie
      name: ta_preview
      description: P01 loopback-only HttpOnly SameSite=Strict session; bootstrap ticket
        is one-use and expires in 5 minutes.
    PreviewReplaySession:
      type: apiKey
      in: cookie
      name: ta_preview_8767
      description: 'P02.1 isolated HttpOnly SameSite=Strict cookie: ta_preview_<configured
        port>; 8767 by default. Cookies do not isolate by port, so its name must differ
        from P01/P02.'
    PreviewMapSession:
      type: apiKey
      in: cookie
      name: ta_preview_8768
      description: P03 isolated loopback session; ta_preview_<configured port>, default
        8768.